requirements.txt
cmapy==0.6.6
einops==0.3.0
imageio==2.9.0
matplotlib==3.7.5
open3d
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pandas==2.0.3
pillow==10.2.0
plotly==5.24.1
plyfile==1.0.3
pykeops==2.2.3
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
tensorboardX==2.6.2.2
tqdm==4.66.1
trimesh==4.3.2
typing_extensions==4.9.0
wandb